MQTT/requirements.txt
paho-mqtt
opcua
